윤성역/구직자페이지 이력서란 끝/ 04-01 17:41

+++ alli/src/main/java/com/smhrd/model/resumeDAO.java
@@ -36,7 +36,7 @@ public class resumeDAO {
 		 return user_resume_list ;
 	 }
 
-	public int deleteResume(BigDecimal deleteIndex) {
+	public int deleteResume(int deleteIndex) {
 		SqlSession sqlSession = sqlSessionFactory.openSession(true);
 		int cnt = sqlSession.delete("resumeDelete",deleteIndex);
 		sqlSession.close();
